SafeStep/requirements.txt
# Safe-Step Flask Application Requirements

# Core Flask and Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7

# Environment Configuration
python-dotenv==1.0.0

# Supabase Integration
supabase==2.17.0

# Database Support
SQLAlchemy==2.0.21

# Form Handling and Validation
Flask-WTF==1.1.1
WTForms==3.0.1

# Email Support
Flask-Mail==0.9.1

# Security and Authentication
bcrypt==4.0.1
PyJWT==2.8.0

# API and JSON Support
Flask-RESTful==0.3.10
marshmallow==3.20.1
Flask-Marshmallow==0.15.0

# Environment Configuration
python-dotenv==1.0.0

# Date and Time Handling
python-dateutil==2.8.2

# HTTP Requests
requests==2.31.0

# Data Processing
pandas==2.1.1
numpy==1.25.2

# Cloud Database Connectors
# For Google Cloud SQL
google-cloud-sql-connector==1.4.1

# For AWS RDS
boto3==1.28.62

# For Azure Database
azure-storage-blob==12.18.3

# Machine Learning (for predictions)
scikit-learn==1.3.0
joblib==1.3.2

# Monitoring and Logging
Flask-Limiter==3.5.0
python-json-logger==2.0.7

# Development and Testing
pytest==7.4.2
pytest-flask==1.2.0
coverage==7.3.2

# Production Server
gunicorn==21.2.0

# Caching
Flask-Caching==2.1.0
redis==4.6.0

# Task Queue (for background jobs)
celery==5.3.2

# Static File Compression
Flask-Compress==1.13